requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
httpx==0.27.2
loguru==0.7.2
numpy==2.1.3
pandas==2.2.3
feedparser==6.0.11
prometheus-client==0.20.0
